fix: update easyeffects service to actually work
easyeffects works as a pseudo service by merely hiding its gui. It
always needs a gui to operate, which means that the full desktop gui
needs to be operational for it to hook into and run. Despite setting it
as late into systemd as I can (with my current knowledge), it still
fails to run on the first attempt.

As a bandage fix, a restart delay was added to give the desktop some
time to finish initializing before trying again. This also still allows
restarts to work.